Crate rome_css_syntax

Modules

kind

Generated file, do not edit by hand, see xtask/codegen

macros

Generated file, do not edit by hand, see xtask/codegen

Macros

T

Utility macro for creating a SyntaxKind through simple macro syntax

map_syntax_node

Reconstruct an AstNode from a SyntaxNode

Structs

CssAnyFunction

CssAnyFunctionFields

CssAnySelectorPatternList

CssAtKeyframes

CssAtKeyframesBody

CssAtKeyframesBodyFields

CssAtKeyframesFields

CssAtKeyframesItemList

CssAtMedia

CssAtMediaFields

CssAtMediaQuery

CssAtMediaQueryConsequent

CssAtMediaQueryConsequentFields

CssAtMediaQueryFeature

CssAtMediaQueryFeatureBoolean

CssAtMediaQueryFeatureBooleanFields

CssAtMediaQueryFeatureCompare

CssAtMediaQueryFeatureCompareFields

CssAtMediaQueryFeatureFields

CssAtMediaQueryFeaturePlain

CssAtMediaQueryFeaturePlainFields

CssAtMediaQueryFeatureRange

CssAtMediaQueryFeatureRangeFields

CssAtMediaQueryFields

CssAtMediaQueryList

CssAtMediaQueryRange

CssAtMediaQueryRangeFields

CssAttribute

CssAttributeFields

CssAttributeList

CssAttributeMatcher

CssAttributeMatcherFields

CssAttributeMeta

CssAttributeMetaFields

CssAttributeModifier

CssAttributeModifierFields

CssAttributeName

CssAttributeNameFields

CssAttributeSelectorPattern

CssAttributeSelectorPatternFields

CssBlock

CssBlockFields

CssClassSelectorPattern

CssClassSelectorPatternFields

CssCombinatorSelectorPattern

CssCombinatorSelectorPatternFields

CssCustomProperty

CssCustomPropertyFields

CssDeclaration

CssDeclarationFields

CssDeclarationImportant

CssDeclarationImportantFields

CssDeclarationList

CssDimension

CssDimensionFields

CssIdSelectorPattern

CssIdSelectorPatternFields

CssIdentifier

CssIdentifierFields

CssKeyframesBlock

CssKeyframesBlockFields

CssKeyframesSelector

CssKeyframesSelectorFields

CssKeyframesSelectorList

CssLanguage

CssNumber

CssNumberFields

CssParameter

CssParameterFields

CssParameterList

CssPercentage

CssPercentageFields

CssPseudoClassSelectorPattern

CssPseudoClassSelectorPatternFields

CssPseudoClassSelectorPatternParameters

CssPseudoClassSelectorPatternParametersFields

CssRatio

CssRatioFields

CssRoot

CssRule

CssRuleFields

CssSelector

CssSelectorFields

CssSelectorList

CssSimpleFunction

CssSimpleFunctionFields

CssString

CssStringFields

CssTypeSelectorPattern

CssTypeSelectorPatternFields

CssUniversalSelectorPattern

CssUniversalSelectorPatternFields

CssUnknown

CssVarFunction

CssVarFunctionFields

CssVarFunctionValue

CssVarFunctionValueFields

DebugSyntaxElementChildren

SyntaxNodeText

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

Enums

CssAnyAtMediaQueryFeatureType

CssAnyAtMediaQueryType

CssAnyAtRule

CssAnyRule

CssAnySelectorPattern

CssAnyValue

CssSyntaxKind

The kind of syntax node, e.g. IDENT, FUNCTION_KW, or FOR_STMT.

TokenAtOffset

There might be zero, one or two leaves at a given offset.

TriviaPieceKind

WalkEvent

WalkEvent describes tree walking process.

Traits

TextLen

Primitives with a textual length that can be passed to TextSize::of.

Type Definitions

CssSyntaxElement

CssSyntaxElementChildren

CssSyntaxList

CssSyntaxNode

CssSyntaxNodeChildren

CssSyntaxToken